+++ CHANGELOG.md
@@ -8,11 +8,22 @@ See also https://docs.resim.ai/changelog/ for all ReSim changes
 
 Changes in this section will be included in the next release.
 
-### v0.13.0 - April 7 2025
+### v0.14.0 - April 8 2025
 
 #### Changed
 
 - It is now possible to pass `pool-labels` to the `resim ingest` command to support log ingestion via the ReSim agent.
+
+
+### v0.13.0 - April 7 2025
+
+#### Added
+
+- The `logs download` command now supports downloading a single log by providing the log name.
+
+#### Changed
+
+- The `logs` commands now support batch logs.
 
 ### v0.12.0 - March 25 2025